+++ src/IceRay/material/pattern/noise/perlin.hpp
@@ -72,9 +72,9 @@
            private:
              static T_scalar Fs_f( T_coord const& P_coord )
               {
-               int x = floor( P_coord[0] );
-               int y = floor( P_coord[1] );
-               int z = floor( P_coord[2] );
+               int x = (int)floor( P_coord[0] );
+               int y = (int)floor( P_coord[1] );
+               int z = (int)floor( P_coord[2] );
 
                T_scalar tx = P_coord[0] - x;
                T_scalar ty = P_coord[1] - y;
